+++ CHANGELOG.md
@@ -1,6 +1,10 @@
-# v1.0.0 / 2018-06-01
+# v1.0.0 / 2018-06-01 - 2018-06-12
 
 - Initial Release
 - Multilingual Support
 - Routes versioning
 - Routes Documentation
+- Use latest majifix common, jurisdiction, priority and service group dependencies
+- Add flags schema to track common service flags
+- Add open311 plugin
+- Expose open311 api endpoint